Seperated the packet structs from the lidar.hpp to lidar_packets.hpp
